refactor: 继续使用宏降低冗余(top1 common)
